requirement.txt
fastapi
uvicorn
httpx
pydantic
pyyaml